+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 * [#86](https://github.com/antoinezanardi/werewolves-assistant-api/issues/86) - Extend player limit to 40.
 * [#87](https://github.com/antoinezanardi/werewolves-assistant-api/issues/87) - Add isRevealed to role structure.
+* [#95](https://github.com/antoinezanardi/werewolves-assistant-api/issues/95) - Limit game history length.
 
 ### ♻️ Refactoring